requirements.txt
rich
tqdm
xarray
pillow
fastapi
uvicorn
rasterio
streamlit
rioxarray
ipykernel
matplotlib
tensorflow
scikit-learn
scikit-image
pystac_client
fastapi[standard]
planetary_computer